+++ R/ident.R
@@ -26,7 +26,6 @@ ident <- function(...) {
   x <- c_character(...)
   structure(x, class = c("ident", "character"))
 }
-setOldClass(c("ident", "character"), ident())
 
 #' @export
 #' @rdname ident
@@ -34,7 +33,6 @@ ident_q <- function(...) {
   x <- c_character(...)
   structure(x, class = c("ident_q", "ident", "character"))
 }
-# setOldClass(c("ident_q", "ident", "character"), ident_q())
 
 #' @export
 print.ident <- function(x, ...) cat(format(x, ...), sep = "\n")